+++ src/test/java/nu/aron/nextbuildnumber/IncrementableTest.java
@@ -47,4 +47,10 @@ class IncrementableTest implements Incrementable {
         assertEquals(0, countDots(""));
     }
 
+    @Test
+    void manuallyBumped() {
+        assertEquals("1.1.0", manuallyBumped("1.1.0-SNAPSHOT", "1.0.555"));
+        assertEquals("1.3.555", manuallyBumped("1.1.0-SNAPSHOT", "1.3.555"));
+        assertEquals("1.1.0", manuallyBumped("1.1.0", "1.0.555"));
+    }
 }