+++ CHANGELOG.md
@@ -1,16 +1,23 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-## [10] - 2020-02-15
+## [17] - 2020-04-13
+
 ### Added
+
+- Tries to detect if version is bumped manually and increments and uses that instead of remote.
+
+## [10] - 2020-02-15
+
 - Tries to be "CI build aware"
 - Takes no action if goals contains "deploy:deploy"
 - Adds branch name to version string if on a branch
 
 ## [8] - 2020-02-07
-### Added
+
 - Now adds a property nextversion.commit to the deployed pom